Revert "Extend the mDNS package with a native extension used on Mac OS"
This reverts commit 7b02aeda7e019a7daa08b2002be956150aaae4e2.

TBR=ricow@google.com
BUG=

Review URL: https://codereview.chromium.org/1413713011 .
