--- pkg/mdns/lib/native/libmdns_extension_lib.dylib.sha1
@@ -1,1 +1,0 @@
-22e12795d48159983c756e96b4355ca63b020676
