--- pkg/mdns/lib/native/libmdns_extension_lib.so.sha1
@@ -1,1 +1,0 @@
-a1bcfd86ed70bf854de519d29cf354871459d825
