+++ pkg/mdns/test/decode_test.dart
@@ -1,6 +1,6 @@
-// Copyright (c) 2015, the Fletch project authors. Please see the AUTHORS file
+// Copyright (c) 2015, the Dart project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
-// BSD-style license that can be found in the LICENSE.md file.
+// BSD-style license that can be found in the LICENSE file.
 
 import 'package:expect/expect.dart';
 import 'package:mdns/src/packet.dart';
@@ -114,3 +114,4 @@ void hexDumpList(List package) {
   }
   if (s.length != 0) print(s);
 }
+
